Extract spectrum.xpub module for Page Objects
