+++ spectrum/input.py
@@ -7,6 +7,7 @@ from pollute import modified_environ
 import mechanicalsoup
 from spectrum import aws, logger
 from spectrum.config import SETTINGS
+from spectrum.xpub import XpubJavaScriptSession
 
 LOGGER = logger.logger(__name__)
 
@@ -237,38 +238,6 @@ class JournalJavaScriptSession:
         selenium_title_smoke_test('xpub', self._driver)
         # expand: click on login button, log in, and check final destination
         return XpubJavaScriptSession(self._driver)
-
-
-class XpubJavaScriptSession:
-    CSS_LOGIN_BUTTON = 'button[data-test-id="login"]'
-    CSS_PROFILE_MENU = 'button[data-test-id="profile-menu"]'
-    CSS_NEW_SUBMISSION_BUTTON = 'button[data-test-id="desktop-new-submission"]'
-
-    def __init__(self, driver):
-        self._driver = driver
-
-    def _log(self, message, *args, **kwargs):
-        LOGGER.info(message, extra={'app':'elife-xpub'}, *args, **kwargs)
-
-    def login(self):
-        login_button = self._driver.find_element_by_css_selector(self.CSS_LOGIN_BUTTON)
-        self._log("Found login button %s `%s`", self.CSS_LOGIN_BUTTON, login_button.text)
-        login_button.click()
-        self._log("Clicked login button %s", self.CSS_LOGIN_BUTTON)
-        profile_menu = self._driver.find_element_by_css_selector(self.CSS_PROFILE_MENU)
-        self._log("Found profile menu %s", self.CSS_PROFILE_MENU)
-        profile_menu.click()
-        self._log("Clicked profile menu %s", self.CSS_PROFILE_MENU)
-
-    def dashboard(self):
-        dashboard_button = self._driver.find_element_by_link_text('Dashboard')
-        dashboard_button.click()
-        new_submission_button = self._driver.find_element_by_css_selector(self.CSS_NEW_SUBMISSION_BUTTON)
-        new_submission_button.click()
-        first_name = self._driver.find_element_by_css_selector('input[name="author.firstName"]')
-        first_name.send_keys("Josiah")
-        last_name = self._driver.find_element_by_css_selector('input[name="author.lastName"]')
-        last_name.send_keys("Carberry")
 
 
 class JournalHtmlSession:
