+++ spectrum/test_xpub.py
@@ -9,4 +9,5 @@ def test_initial_submission(get_selenium_driver):
     journal_session = input.JOURNAL.javascript_session(get_selenium_driver())
     xpub_session = journal_session.submit()
     xpub_session.login()
-    xpub_session.dashboard()
+    author_page = xpub_session.dashboard().create_initial_submission()
+    author_page.next()
